# trade-mobile/src/main/java/com/centaline/parportal/mobile/chat/web/ContactController.java
package com.centaline.parportal.mobile.chat.web;

import java.util.HashMap;
import java.util.List;
import java.util.Map;

import javax.annotation.Resource;

import org.springframework.data.domain.Page;
import org.springframework.stereotype.Controller;
import org.springframework.web.bind.annotation.RequestMapping;
import org.springframework.web.bind.annotation.RequestMethod;
import org.springframework.web.bind.annotation.RequestParam;
import org.springframework.web.bind.annotation.ResponseBody;

import com.aist.common.quickQuery.bo.JQGridParam;
import com.aist.common.quickQuery.service.QuickGridService;
import com.alibaba.fastjson.JSONArray;
import com.alibaba.fastjson.JSONObject;

@Controller
@RequestMapping("/chat")
public class ContactController {
	@Resource(name = "quickGridService")
	private QuickGridService quickGridService;

	@ResponseBody
	@RequestMapping(value="casecontact")
	public JSONObject casecontact(@RequestParam(required=true) String caseCode) {
		JQGridParam gp = new JQGridParam();
		gp.setPagination(false);

		gp.setQueryId("casecontact");

		Map<String, Object> paramter = new HashMap<String, Object>();
		paramter.put("caseCode", caseCode);
		gp.putAll(paramter);

		Page<Map<String, Object>> returnPage = quickGridService.findPageForSqlServer(gp);

		JSONObject result = new JSONObject();

		result.put("total", returnPage.getTotalPages());
		// result.put("records", returnPage.getTotalElements());
		List<Map<String, Object>> contentList = returnPage.getContent();

		/*
		 * JSONArray content = new JSONArray(); for (int i = 0; i <
		 * contentList.size(); i++) { Map<String, Object> mapObj =
		 * contentList.get(i);
		 * 
		 * content.add(JSONObject.toJSON(mapObj)); }
		 */
		result.put("rows", contentList);

		return result;
	}

	@ResponseBody
	@RequestMapping(value="contact")
	public JSONObject contact(@RequestParam(required=true)Integer page, @RequestParam(required=true)Integer pageSize, String sidx, String sord,@RequestParam(required=true) String q_text) {

		JQGridParam gp = new JQGridParam();
		gp.setPagination(true);
		gp.setPage(page);
		gp.setRows(pageSize);
		gp.setSidx(sidx);
		gp.setSord(sord);
		gp.setQueryId("contact");

		Map<String, Object> paramter = new HashMap<String, Object>();
		paramter.put("q_text", q_text);
		gp.putAll(paramter);

		Page<Map<String, Object>> returnPage = quickGridService.findPageForSqlServer(gp);

		JSONObject result = new JSONObject();

		result.put("total", returnPage.getTotalPages());
		result.put("page", page);
		result.put("total", returnPage.getTotalPages());
		result.put("records", returnPage.getTotalElements());
		result.put("pageSize", pageSize);
		List<Map<String, Object>> contentList = returnPage.getContent();

		result.put("rows", contentList);

		return result;
	}

}