# trade-mobile/swagger/api/swagger/swagger.yaml
swagger: "2.0"
info:
  version: "0.0.1"
  title: Hello World App
# during dev, should point to your local machine
host: swagger.centaline.com:8086
...
basePath: /
...
schemes:
  # tip: remove http to make production-grade
  - http
  - https
# format of bodies a client can send (Content-Type)
consumes:
  - application/json
# format of the responses to the client (Accepts)
produces:
  - application/json
paths:
...
  /chat/contact:
    # binds a127 app logic to a route
    x-swagger-router-controller: defult_response
    get:
      description: demo config
      # used as the method name of the controller
      operationId: contactRsponse
      parameters:
        - name: token
          in: query
          description: token
          required: true
          type: string
        - name: page
          in: query
          description: page
          required: true
          type: integer
        - name: pageSize
          in: query
          description: pageSize
          required: true
          type: integer 
        - name: sidx
          in: query
          description: sidx
          required: true
          type: string
        - name: sord
          in: query
          description: sord
          required: false
          type: string  
        - name: q_text
          in: query
          description: q_text
          required: false
          type: string 
      responses:
        "200":
          description: Success
          schema:
            # a pointer to a definition
            $ref: "#/definitions/getContactSuccessResponse"
  /chat/casecontact:
    # binds a127 app logic to a route
    x-swagger-router-controller: defult_response
    post:
      description: demo config
      # used as the method name of the controller
      operationId: casecontactRsponse
      parameters:
        - name: token
          in: query
          description: token
          required: true
          type: string
        - name: caseCode
          in: query
          description: token
          required: true
          type: string
      responses:
       "200":
          description: Success
          schema:
            # a pointer to a definition
            $ref: "#/definitions/getCasecontactSuccessResponse"
...
definitions:
...
  contactUser:
    type: object
    properties:
      org: 
        type: string
      avatarUrl:
        type: string
      name:
        type: string
      mobile: 
        type: string
      userName:
        type: string
      job: 
        type: string
      userId: 
        type: string
  getContactSuccessResponse:
    required:
      - isSuccess
    properties:
      isSuccess:
        type: boolean
      msg:
        type: string
      empId:
        type: string
      token:
        type: string
      content:
       $ref: "#/definitions/contactContent"
  contactContent:
    properties:
      total:
        type: integer
      records:
        type: integer
      pageSize:
        type: integer
      page:
        type: integer
      rows:
        type: array
        items:
          $ref: '#/definitions/contactUser'
  getCasecontactSuccessResponse:
    required:
      - isSuccess
    properties:
      isSuccess:
        type: boolean
      msg:
        type: string
      empId:
        type: string
      token:
        type: string
      content:
        $ref: "#/definitions/casecontactContent"
  casecontactContent:
     properties:
      total:
        type: integer
      rows:
        type: array
        items:
          $ref: '#/definitions/contactUser'
...